+++ app/(main)/(campaigns)/campaigns.tsx
@@ -13,6 +13,7 @@ import CampaignsFilledState from "@/components/campaigns/CampaignsFilledState";
 import { TextInput } from "react-native-paper";
 import { useCampaignContext } from "@/contexts/campaign-context.provider";
 import { Campaign } from "@/types/campaign";
+import { useOrganizationContext } from "@/contexts";
 
 const Campaigns = () => {
   const [searchQuery, setSearchQuery] = useState("");
@@ -23,6 +24,13 @@ const Campaigns = () => {
     campaigns,
     getCampaigns,
   } = useCampaignContext();
+  const {
+    currentOrganization,
+  } = useOrganizationContext();
+
+  const fetchCampaigns = async () => {
+    getCampaigns();
+  }
 
   const handleSearch = (query: string) => {
     setSearchQuery(query);
@@ -37,8 +45,16 @@ const Campaigns = () => {
   };
 
   useEffect(() => {
-    getCampaigns();
-  }, []);
+    if (campaigns) {
+      setFilteredCampaigns(campaigns);
+    }
+  }, [campaigns]);
+
+  useEffect(() => {
+    if (currentOrganization) {
+      fetchCampaigns();
+    }
+  }, [currentOrganization]);
 
   if (!campaigns) {
     return null;
